🛠️ src/pages/Blog.tsx -> Refactored to use hooks 🛠️ src/pages/BlogPost.tsx -> Simplified post fetching
