+++ src/pages/Blog.tsx
@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import BlogContainer from '@/components/blog/BlogContainer';
 import BlogHeader from '@/components/blog/BlogHeader';
-import BlogSection from '@/components/blog/BlogSection';
-import { type BlogPost, initialBlogPosts } from '@/data/blogData';
+import BlogContent from '@/components/blog/BlogContent';
+import useBlogPosts from '@/hooks/useBlogPosts';
 
 /**
  * Blog page component that displays a list of blog posts
@@ -9,26 +10,13 @@ import { type BlogPost, initialBlogPosts } from '@/data/blogData';
  * Implements loading states and error handling
  */
 const BlogPage: React.FC = () => {
-  const [posts, setPosts] = useState<BlogPost[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    // Simulate loading delay
-    const timer = setTimeout(() => {
-      setPosts(initialBlogPosts);
-      setIsLoading(false);
-    }, 500);
-
-    return () => clearTimeout(timer);
-  }, []);
+  const { posts, isLoading } = useBlogPosts();
 
   return (
-    <div className="min-h-screen bg-[#FFFFFF]">
+    <BlogContainer>
       <BlogHeader />
-      <div className="container py-16">
-        <BlogSection posts={posts} isLoading={isLoading} />
-      </div>
-    </div>
+      <BlogContent posts={posts} isLoading={isLoading} />
+    </BlogContainer>
   );
 };
